Rename to ValidityPeriodUtil
