+++ administration/src/util/validityPeriod.test.ts
@@ -1,5 +1,5 @@
 import { format } from 'date-fns'
-import { dateToDaysSinceEpoch, daysSinceEpochToDate } from './day'
+import { dateToDaysSinceEpoch, daysSinceEpochToDate } from './validityPeriod'
 
 describe('dateToDaysSinceEpoch', () => {
   test('epoch', () => {
